Create Submit jobs to a list of agents
Script that reads an external file for a listing of agents.  Loops through the list and submits "Sleep120" job for each agent in the list.  The list is stored in an external text file.  The file must contain agents in a column of strings.  For example:

Agent1
Agent2
Agent3
etc...