FTPFilesOnlyForEachFolder.ps1
Powershell script to use JAMS to download all files with a specific file mask to a destination folder up to 3 folders deep for each folder in the root of the FTP directory. The script will pick out the specified files in each folder only, it will not download the folders themselves.